# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAndtAdmin",
    "description": "Multi-tenant Enterprise Management System API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/menus": {
      "get": {
        "tags": [
          "Menus"
        ],
        "summary": "List Menus",
        "description": "List all menus with pagination and filters.\n\nRequires menu:list permission.",
        "operationId": "list_menus_api_v1_menus_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Keyword"
            }
          },
          {
            "name": "menu_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Menu Type"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Menus"
        ],
        "summary": "Create Menu",
        "description": "Create a new menu.",
        "operationId": "create_menu_api_v1_menus_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MenuCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Create Menu Api V1 Menus Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/menus/tree/all": {
      "get": {
        "tags": [
          "Menus"
        ],
        "summary": "Get Menu Tree",
        "description": "Get complete menu tree (without permission filtering).\n\nFor admin use. Requires menu:list permission.",
        "operationId": "get_menu_tree_api_v1_menus_tree_all_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/menus/user": {
      "get": {
        "tags": [
          "Menus"
        ],
        "summary": "Get User Menu Tree",
        "description": "Get current user's menu tree.\n\nReturns menu tree based on user's permissions.",
        "operationId": "get_user_menu_tree_api_v1_menus_user_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MenuTreeNode"
                  },
                  "type": "array",
                  "title": "Response Get User Menu Tree Api V1 Menus User Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/menus/{menu_id}": {
      "get": {
        "tags": [
          "Menus"
        ],
        "summary": "Get Menu",
        "description": "Get menu by ID.\n\nRequires menu:query permission.",
        "operationId": "get_menu_api_v1_menus__menu_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "menu_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Menu Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Menus"
        ],
        "summary": "Update Menu",
        "description": "Update a menu.",
        "operationId": "update_menu_api_v1_menus__menu_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "menu_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Menu Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MenuUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Update Menu Api V1 Menus  Menu Id  Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Menus"
        ],
        "summary": "Delete Menu",
        "description": "Delete a menu.",
        "operationId": "delete_menu_api_v1_menus__menu_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "menu_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Menu Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Delete Menu Api V1 Menus  Menu Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MenuCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "菜单名称"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "菜单标题"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id",
            "description": "父菜单ID"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path",
            "description": "路由路径"
          },
          "component": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Component",
            "description": "组件路径"
          },
          "redirect": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Redirect",
            "description": "重定向路径"
          },
          "icon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon",
            "description": "图标"
          },
          "sort": {
            "type": "integer",
            "title": "Sort",
            "description": "排序",
            "default": 0
          },
          "type": {
            "type": "integer",
            "title": "Type",
            "description": "类型:1目录,2菜单,3按钮",
            "default": 1
          },
          "permission_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Permission Code",
            "description": "权限标识"
          },
          "status": {
            "type": "integer",
            "title": "Status",
            "description": "状态:0禁用,1启用",
            "default": 1
          },
          "visible": {
            "type": "integer",
            "title": "Visible",
            "description": "是否可见:0隐藏,1显示",
            "default": 1
          },
          "is_cache": {
            "type": "integer",
            "title": "Is Cache",
            "description": "是否缓存:0否,1是",
            "default": 0
          },
          "is_external": {
            "type": "integer",
            "title": "Is External",
            "description": "是否外链:0否,1是",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "name",
          "title"
        ],
        "title": "MenuCreate",
        "description": "Schema for creating menu."
      },
      "MenuTreeNode": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "菜单名称"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "菜单标题"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id",
            "description": "父菜单ID"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path",
            "description": "路由路径"
          },
          "component": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Component",
            "description": "组件路径"
          },
          "redirect": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Redirect",
            "description": "重定向路径"
          },
          "icon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon",
            "description": "图标"
          },
          "sort": {
            "type": "integer",
            "title": "Sort",
            "description": "排序",
            "default": 0
          },
          "type": {
            "type": "integer",
            "title": "Type",
            "description": "类型:1目录,2菜单,3按钮",
            "default": 1
          },
          "permission_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Permission Code",
            "description": "权限标识"
          },
          "status": {
            "type": "integer",
            "title": "Status",
            "description": "状态:0禁用,1启用",
            "default": 1
          },
          "visible": {
            "type": "integer",
            "title": "Visible",
            "description": "是否可见:0隐藏,1显示",
            "default": 1
          },
          "is_cache": {
            "type": "integer",
            "title": "Is Cache",
            "description": "是否缓存:0否,1是",
            "default": 0
          },
          "is_external": {
            "type": "integer",
            "title": "Is External",
            "description": "是否外链:0否,1是",
            "default": 0
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "tenant_id": {
            "type": "integer",
            "title": "Tenant Id"
          },
          "children": {
            "items": {
              "$ref": "#/components/schemas/MenuTreeNode"
            },
            "type": "array",
            "title": "Children",
            "description": "子菜单"
          }
        },
        "type": "object",
        "required": [
          "name",
          "title",
          "id",
          "tenant_id"
        ],
        "title": "MenuTreeNode",
        "description": "Menu tree node with children."
      },
      "MenuUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          },
          "component": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Component"
          },
          "redirect": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Redirect"
          },
          "icon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon"
          },
          "sort": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sort"
          },
          "type": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          },
          "permission_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Permission Code"
          },
          "status": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "visible": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Visible"
          },
          "is_cache": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Cache"
          },
          "is_external": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is External"
          }
        },
        "type": "object",
        "title": "MenuUpdate",
        "description": "Schema for updating menu."
      },
      "Response": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "响应码",
            "default": 200
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "响应消息",
            "default": "success"
          },
          "timestamp": {
            "type": "integer",
            "title": "Timestamp",
            "description": "时间戳"
          },
          "data": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "响应数据"
          }
        },
        "type": "object",
        "required": [
          "timestamp"
        ],
        "title": "Response",
        "description": "Standard response model."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}